Hoàn thành quản lý người dùng (thêm, sửa, xóa, hiển thị)

+++ Model/DAO/UserDAO.cs
@@ -2,6 +2,8 @@
 using System.Linq;
 using System.Collections.Generic;
 using System.Data.Entity;
+using System;
+using Common;
 
 namespace Model.DAO
 {
@@ -40,6 +42,8 @@ namespace Model.DAO
         {
             try
             {
+                user.Password = Encryptor.MD5Hash(user.Password);
+                user.CreatedDate = DateTime.Now;
                 db.Users.Add(user);
                 db.SaveChanges();
                 return true;
@@ -52,19 +56,35 @@ namespace Model.DAO
             return db.Users.Find(id);
         }
 
+        public bool CheckUsernameIsExist(string username)
+        {
+            return db.Users.Count(x => x.UserName == username) > 0;
+        }
+
+        public bool CheckEmailIsExist(string email)
+        {
+            return db.Users.Count(x => x.Email == email) > 0;
+        }
+
         public bool Update(User user)
         {
+            User userChange = db.Users.SingleOrDefault(x => x.Id == user.Id);
             try
             {
-                db.Entry(user).State = EntityState.Modified;
-                db.SaveChanges();
-                return true;
-            }
-            catch
-            {
+                if (userChange != null)
+                {
+                    userChange.FullName = user.FullName;
+                    userChange.Address = user.Address;
+                    userChange.Email = user.Email;
+                    userChange.Phone = user.Phone;
+                    userChange.Status = user.Status;
+                    db.SaveChanges();
+                    return true;
+                }
                 return false;
             }
-
+            catch { return false; }
+            
         }
 
         public bool Delete(string id)
@@ -72,7 +92,7 @@ namespace Model.DAO
             try
             {
                 var user = db.Users.Find(int.Parse(id));
-                if (user != null)
+                if (user != null && user.UserName != "admin")
                 {
                     db.Users.Remove(user);
                     db.SaveChanges();
